Added target suppor (NOT TESTED)
- Fixed missing if statement for the ped deletion
- Optimized distance check

+++ client.lua
@@ -1,3 +1,4 @@
+local Target = 'None'; -- ox_target or qtarget or qb-target
 local Monkeys = {
   [1] = {
       coords = vec3(-1348.3923, -1446.5963, 4.4976),
@@ -55,13 +56,43 @@ function Draw3DText(x, y, z, text)
 end
 
 CreateThread(function()
+  if Target ~= 'None' then
+    for i = 1, #Monkeys do
+      local monkey = Monkeys[i];
+      if Target == 'qb-target' or Target == 'qtarget' then
+        exports[Target]:AddBoxZone('ec_moneky_number_'..i, monkey.coords, 1, 1, {
+          name = 'ec_moneky_number_'..i,
+          heading = monkey.heading,
+          debugPoly = false,
+          minZ = monkey.coords.z - 1,
+          maxZ = monkey.coords.z + 1
+        }, {
+          options = {
+            {
+              label = 'Interact',
+              icon = 'fas fa-hand',
+              action = function()
+                SetPedArmour(PlayerPedId(), 100)
+                local maxHealth = GetEntityMaxHealth(PlayerPedId());
+                SetEntityHealth(PlayerPedId(), maxHealth);
+              end
+            }
+          },
+          distance = 1.5
+        });
+      end
+    end
+  else
+    local waittime = 500;
     while true do
-        for i = 1, #Monkeys do
-          local monkey = Monkeys[i];
-          local coords = monkey.coords;
-          local distance = #(coords - GetEntityCoords(PlayerPedId()));
-          if distance <= 10 then
-            Draw3DText(coords.x, coords.y, coords.z, '[E] To maxHealth');
+      for i = 1, #Monkeys do
+        local monkey = Monkeys[i];
+        local coords = monkey.coords;
+        local distance = #(coords - GetEntityCoords(PlayerPedId()));
+        if distance <= 10 then
+          waittime = 3;
+          Draw3DText(coords.x, coords.y, coords.z, '[E] To maxHealth');
+          if distance <= 3 then
             if IsControlJustPressed(0, 38) then
               SetPedArmour(PlayerPedId(), 100)
               local maxHealth = GetEntityMaxHealth(PlayerPedId());
@@ -69,14 +100,21 @@ CreateThread(function()
             end
           end
           break;
+        else
+          waittime = 500;
         end
-        Wait(3);
+      end
+      Wait(waittime);
     end
+  end
 end)
 
 AddEventHandler('onResourceStop', function(rsc)
   if rsc == GetCurrentResourceName() then
     for i = 1, #Monkeys do
+      if not DoesEntityExist(PlayerPedId()) then
+        return;
+      end
       DeleteEntity(Monkeys[i].ped);
     end
   end
